Conclusão dos testes

+++ src/main/java/com/poveda/training_bank/clientes/view/ClientesEndpoint.java
@@ -49,14 +49,14 @@ public class ClientesEndpoint {
     }
 
    @GetMapping(path = "/clientes/{idCliente}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ClienteResource findById(@PathVariable String idCliente){
+    public ClienteResource findById(@PathVariable Long idCliente){
         System.out.println("Busca por ID");
 
-       if(StringUtils.isBlank(idCliente) || idCliente == null) throw new ValidationException("Solicitacao invalida");
+       if(idCliente == null) throw new ValidationException("Solicitacao invalida");
 
         final Cliente clienteResource = service.findById(idCliente);
 
-        if(clienteResource == null) throw new NotFoundException("Conta nao localizada, id " + idCliente);
+        if(clienteResource == null) throw new NotFoundException("Cliente inexistente, id " + idCliente);
 
         return clienteResourceAssembler.toResource(clienteResource);
 
@@ -64,19 +64,23 @@ public class ClientesEndpoint {
     }
 
     @GetMapping(path = "/clientes/{idCliente}/contas", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ContaResource findConta(@PathVariable String idCliente){
+    public ContaResource findConta(@PathVariable Long idCliente){
        System.out.println("Busca por conta");
 
-       final Cliente cliente = service.findById(idCliente);
-       Conta conta = cliente.getConta();
+       if(idCliente == null) throw new ValidationException("Conta invalida");
 
+       final Cliente cliente = service.findById(idCliente);
+
+       if(cliente == null) throw new NotFoundException("Conta inexistente");
+
+       Conta conta = cliente.getConta();
 
 
        return clienteResourceAssembler.toResource(cliente, conta);
     }
 
     @GetMapping(path = "/clientes/{idCliente}/contas/cards", produces = MediaType.APPLICATION_JSON_VALUE)
-    public CardResource findCard(@PathVariable String idCliente){
+    public CardResource findCard(@PathVariable Long idCliente){
        System.out.println("Busca Cartões");
 
        final Cliente cliente = service.findById(idCliente);
@@ -88,13 +92,15 @@ public class ClientesEndpoint {
 
     @DeleteMapping(value = "/clientes/{nome}")
     public void delete(@PathVariable String nome){
+        if(StringUtils.isBlank(nome)) throw new ValidationException("Cliente invalido");
         service.delete(nome);
         System.out.println("Cliente removido com sucesso");
+
     }
 
     @PutMapping(value = "/clientes/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public Cliente update(@RequestBody Cliente cliente,
-                          @PathVariable String id){
+                          @PathVariable Long id){
         service.update(cliente, id);
         System.out.println("Cliente atualizado com sucesso");
         return cliente;
